requirements.txt
tensorflow==2.15.0
keras==2.15.0
flask==3.0.0
numpy==1.24.3
pillow==10.1.0
werkzeug==3.0.1
opencv-python==4.8.1.78
matplotlib==3.8.2
scikit-learn==1.3.2
pandas==2.1.4
